Ported all remaining plugins to Python 3.x Refactored REST webservice client Fixed code review issues
git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/grid-control/trunk/grid-control@1643 29098800-e21f-4f28-a8ff-84ee76f0ac90

+++ packages/grid_control/parameters/psource_meta.py
@@ -242,7 +242,7 @@ class CrossParameterSource(MultiParameterSource):
 	def fillParameterInfo(self, pNum, result):
 		for (psource, maxN, prev) in self.quickFill:
 			if maxN:
-				psource.fillParameterInfo((pNum / prev) % maxN, result)
+				psource.fillParameterInfo(int((pNum / prev) % maxN), result)
 			else:
 				psource.fillParameterInfo(pNum, result)
